# apps/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flowza API",
    "version": "0.1.0",
    "description": "Документация по уже реализованным backend-эндпоинтам Flowza. Спецификация собрана по текущим NestJS controller и DTO."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local backend"
    }
  ],
...
  "paths": {
...
    "/tenants": {
      "get": {
        "tags": ["Tenants"],
        "summary": "Получить список активных организаций",
        "responses": {
          "200": {
            "description": "Список активных организаций",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicTenantListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tenants"],
        "summary": "Создать организацию",
        "description": "Доступно только для superAdmin.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTenantDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Организация создана",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TenantDetails" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/tenants/accessible": {
      "get": {
        "tags": ["Tenants"],
        "summary": "Получить активные организации, доступные текущему сотруднику",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Доступные организации",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicTenantListItem"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/tenants/manageable": {
      "get": {
        "tags": ["Tenants"],
        "summary": "Получить управляемые организации",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Список управляемых организаций",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManageableOrganizationListItem"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/tenants/{slug}/catalog": {
      "get": {
        "tags": ["Tenants"],
        "summary": "Получить публичный каталог организации по slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "tokyo-roll"
          }
        ],
        "responses": {
          "200": {
            "description": "Публичный каталог",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCatalogResponse"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/tenants/{tenantId}/management": {
      "get": {
        "tags": ["Tenants"],
        "summary": "Получить management view организации",
        "description": "Доступно для superAdmin, admin и moderator в рамках доступных организаций.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Организация, категории и товары",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationManagementResponse"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/tenants/{tenantId}": {
      "patch": {
        "tags": ["Tenants"],
        "summary": "Обновить организацию",
        "description": "Для superAdmin и admin. Поля `isActive` и `subscription` доступны только superAdmin.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTenantDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Организация обновлена",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TenantDetails" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "delete": {
        "tags": ["Tenants"],
        "summary": "Удалить организацию",
        "description": "Доступно только для superAdmin.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Организация удалена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "example": 1 }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Ошибка запроса",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "statusCode": { "type": "integer", "example": 400 },
          "message": {
            "type": "string",
            "example": "Ошибка валидации"
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string", "example": "email" },
                "message": {
                  "type": "string",
                  "example": "Email должен быть валидным адресом"
                }
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-29T18:00:00.000Z"
          },
          "path": { "type": "string", "example": "/auth/register" }
        },
        "required": ["success", "statusCode", "message", "errorCode", "timestamp", "path"]
      },
...
      "PublicTenantListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Tokyo Roll" },
          "slug": { "type": "string", "example": "tokyo-roll" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Японская кухня и роллы"
          }
        },
        "required": ["id", "name", "slug", "description"]
      },
      "ManageableOrganizationListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Tokyo Roll" },
          "slug": { "type": "string", "example": "tokyo-roll" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Японская кухня и роллы"
          },
          "isActive": { "type": "boolean", "example": true }
        },
        "required": ["id", "name", "slug", "description", "isActive"]
      },
      "TenantDetails": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Tokyo Roll" },
          "slug": { "type": "string", "example": "tokyo-roll" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Японская кухня и роллы"
          },
          "heroTitle": {
            "type": "string",
            "nullable": true,
            "example": "Свежие роллы каждый день"
          },
          "heroSubtitle": {
            "type": "string",
            "nullable": true,
            "example": "Доставка за 30 минут"
          },
          "heroDescription": {
            "type": "string",
            "nullable": true,
            "example": "Собираем сеты из свежих ингредиентов"
          },
          "heroImageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/hero.jpg"
          },
          "seoTitle": {
            "type": "string",
            "nullable": true,
            "example": "Tokyo Roll | Доставка роллов"
          },
          "seoDescription": {
            "type": "string",
            "nullable": true,
            "example": "Роллы, суши и сеты с доставкой"
          },
          "isActive": { "type": "boolean", "example": true },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+79990000000"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "Москва"
          },
          "timezone": { "type": "string", "example": "Europe/Moscow" },
          "workingHours": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "example": { "mon": "10:00-22:00", "tue": "10:00-22:00" }
          },
          "deliveryFee": { "type": "integer", "example": 199 },
          "minOrderAmount": { "type": "integer", "example": 1000 },
          "subscription": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-12-31T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "heroTitle",
          "heroSubtitle",
          "heroDescription",
          "heroImageUrl",
          "seoTitle",
          "seoDescription",
          "isActive",
          "phone",
          "address",
          "timezone",
          "workingHours",
          "deliveryFee",
          "minOrderAmount"
        ]
      },
      "PublicCatalogCategory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "name": { "type": "string", "example": "Роллы" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Фирменные роллы"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/category.jpg"
          },
          "sortOrder": { "type": "integer", "example": 1 }
        },
        "required": ["id", "name", "description", "imageUrl", "sortOrder"]
      },
      "PublicCatalogProduct": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 7 },
          "categoryId": { "type": "integer", "example": 5 },
          "name": { "type": "string", "example": "Филадельфия" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Лосось, сыр, рис"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/product.jpg"
          },
          "badgeText": { "type": "string", "nullable": true, "example": "Хит" },
          "price": { "type": "integer", "example": 520 },
          "currency": { "type": "string", "example": "RUB" }
        },
        "required": [
          "id",
          "categoryId",
          "name",
          "description",
          "imageUrl",
          "badgeText",
          "price",
          "currency"
        ]
      },
      "PublicCatalogResponse": {
        "type": "object",
        "properties": {
          "tenant": { "$ref": "#/components/schemas/TenantDetails" },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicCatalogCategory" }
          },
          "products": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicCatalogProduct" }
          }
        },
        "required": ["tenant", "categories", "products"]
      },
      "OrganizationCategory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "tenantId": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Роллы" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Фирменные роллы"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/category.jpg"
          },
          "sortOrder": { "type": "integer", "example": 1 },
          "isActive": { "type": "boolean", "example": true }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "description",
          "imageUrl",
          "sortOrder",
          "isActive"
        ]
      },
      "OrganizationProduct": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 7 },
          "tenantId": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "example": 5 },
          "name": { "type": "string", "example": "Филадельфия" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Лосось, сыр, рис"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/product.jpg"
          },
          "badgeText": { "type": "string", "nullable": true, "example": "Хит" },
          "price": { "type": "integer", "example": 520 },
          "currency": { "type": "string", "example": "RUB" },
          "isActive": { "type": "boolean", "example": true }
        },
        "required": [
          "id",
          "tenantId",
          "categoryId",
          "name",
          "description",
          "imageUrl",
          "badgeText",
          "price",
          "currency",
          "isActive"
        ]
      },
      "OrganizationManagementResponse": {
        "type": "object",
        "properties": {
          "tenant": { "$ref": "#/components/schemas/TenantDetails" },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrganizationCategory" }
          },
          "products": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrganizationProduct" }
          }
        },
        "required": ["tenant", "categories", "products"]
      },
      "CreateTenantDto": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string", "example": "Tokyo Roll" },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "example": "tokyo-roll"
          },
          "description": {
            "type": "string",
            "example": "Японская кухня и роллы"
          },
          "heroTitle": {
            "type": "string",
            "example": "Свежие роллы каждый день"
          },
          "heroSubtitle": {
            "type": "string",
            "example": "Доставка за 30 минут"
          },
          "heroDescription": {
            "type": "string",
            "example": "Сеты, суши и фирменные роллы"
          },
          "heroImageUrl": {
            "type": "string",
            "example": "https://example.com/hero.jpg"
          },
          "seoTitle": {
            "type": "string",
            "example": "Tokyo Roll | Доставка роллов"
          },
          "seoDescription": {
            "type": "string",
            "example": "Роллы, суши и сеты с доставкой"
          },
          "subscription": {
            "type": "string",
            "format": "date-time",
            "example": "2026-12-31T00:00:00.000Z"
          },
          "phone": { "type": "string", "example": "+79990000000" },
          "address": { "type": "string", "example": "Москва" },
          "timezone": { "type": "string", "example": "Europe/Moscow" },
          "workingHours": {
            "type": "object",
            "additionalProperties": true,
            "example": { "mon": "10:00-22:00", "tue": "10:00-22:00" }
          },
          "deliveryFee": { "type": "integer", "example": 199 },
          "minOrderAmount": { "type": "integer", "example": 1000 }
        }
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Tokyo Roll Premium" },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "example": "tokyo-roll"
          },
          "description": {
            "type": "string",
            "example": "Обновленное описание"
          },
          "heroTitle": {
            "type": "string",
            "example": "Новый геройский заголовок"
          },
          "heroSubtitle": {
            "type": "string",
            "example": "Еще быстрее, еще свежее"
          },
          "heroDescription": {
            "type": "string",
            "example": "Расширили ассортимент"
          },
          "heroImageUrl": {
            "type": "string",
            "example": "https://example.com/new-hero.jpg"
          },
          "seoTitle": { "type": "string", "example": "Tokyo Roll Premium" },
          "seoDescription": {
            "type": "string",
            "example": "Новая SEO-описание"
          },
          "isActive": { "type": "boolean", "example": true },
          "subscription": {
            "type": "string",
            "format": "date-time",
            "example": "2026-12-31T00:00:00.000Z"
          },
          "phone": { "type": "string", "example": "+79990000000" },
          "address": { "type": "string", "example": "Москва, ул. Пример, 1" },
          "timezone": { "type": "string", "example": "Europe/Moscow" },
          "workingHours": {
            "type": "object",
            "additionalProperties": true,
            "example": { "fri": "10:00-23:00" }
          },
          "deliveryFee": { "type": "integer", "example": 250 },
          "minOrderAmount": { "type": "integer", "example": 1200 }
        }
      },
...
    }
  }
}